requirements.txt
asgiref==3.5.0
boto3==1.22.9
botocore==1.25.9
Django==4.0.4
django-appconf==1.0.5
django-decouple==2.1
django-imagekit==4.1.0
django-quill-editor==0.1.40
django-storages==1.12.3
gunicorn==20.1.0
jmespath==1.0.0
pilkit==2.0
Pillow==9.1.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
